+++ src/providers.jsx
@@ -2,8 +2,6 @@ import { FullscreenProvider } from '@/libs/fullscreen';
 import {
   GAPI_API_KEY,
   GAPI_CLIENT_ID,
-  GAPI_DRIVE_APPDATA_SCOPE,
-  GAPI_DRIVE_FILE_SCOPE,
   GoogleAPIProvider,
 } from '@/libs/googleapi';
 
@@ -18,10 +16,7 @@ import { VideoCacheProvider } from './recorder/cache';
  */
 export default function Providers({ children }) {
   return (
-    <GoogleAPIProvider
-      apiKey={GAPI_API_KEY}
-      clientId={GAPI_CLIENT_ID}
-      scopes={[GAPI_DRIVE_FILE_SCOPE, GAPI_DRIVE_APPDATA_SCOPE]}>
+    <GoogleAPIProvider apiKey={GAPI_API_KEY} clientId={GAPI_CLIENT_ID}>
       <GoogleDriveSyncProvider>
         <VideoCacheProvider>
           <FullscreenProvider>